moved some fixtures around
